+++ src/main/env.d.ts
@@ -14,8 +14,10 @@ interface ImportMetaEnv {
   readonly VITE_API_ENDPOINT_INFO_SHIPS_VERSION: string
   readonly VITE_API_ENDPOINT_RECORD_DETAIL: string
   readonly VITE_API_ENDPOINT_RECORD_STATS: string
+  readonly VITE_API_ENDPOINT_RECORD_STATS_NEW: string
   readonly VITE_API_ENDPOINT_RECORD_EVALUATE: string
   readonly VITE_API_ENDPOINT_RECORD_CLAN_INFO: string
+  readonly VITE_API_ENDPOINT_RECORD_BATTLE_ENRICH: string
 }
 
 interface ImportMeta {